BIOps_Deployment_Scripts/Asset as Code 1/CodeBuild/Assets_as_Code/Credentialing/requirements.txt
awscli
boto3-1.20.32-py3-none-any.whl
botocore-1.23.32-py3-none-any.whl
pandas